requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
atomicwrites==1.3.0       # via fpyutils
click==7.0
et-xmlfile==1.0.1         # via openpyxl
fpyutils==1.1.0           # via md-toc
jdcal==1.4.1              # via openpyxl
md-toc==6.0.2
openpyxl==3.0.4
pypandoc==1.4
pyyaml-include==1.2
pyyaml==5.3
rtyaml==1.0.0
wheel==0.34.2             # via pypandoc

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
